requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
Pillow==8.3.1
torch==1.9.0
torchvision==0.10.0
